Pull up a method
